refactor: limit 파라미터 타입을 Long에서 int로 변경

+++ src/main/java/roomescape/theme/ThemeService.java
@@ -14,7 +14,7 @@ public class ThemeService {
 
     private static final int POPULAR_PERIOD = 7;
     private static final int POPULAR_OFFSET = 1;
-    private static final long POPULAR_LIMIT = 10L;
+    private static final int POPULAR_LIMIT = 10;
     private final ThemeRepository themeRepository;
 
     public ThemeService(ThemeRepository themeRepository) {
